+++ .env.example
@@ -2,7 +2,8 @@ PORT=4000
 POSTGRES_DB=home_library
 POSTGRES_USER=postgres
 POSTGRES_PASSWORD=postgres
-DATABASE_URL=postgres://postgres:postgres@postgres:5432/home_library
+POSTGRES_PORT=5432
+DATABASE_URL=postgres://${POSTGRES_USER}:${POSTGRES_PASSWORD}@localhost:${POSTGRES_PORT}/${POSTGRES_DB}
 
 CRYPT_SALT=10
 JWT_SECRET_KEY=secret123123
