+++ README.md
@@ -1,11 +1,13 @@
 # Home Library Service
 
-Task: https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/rest-service/assignment.md
+Task1: https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/rest-service/assignment.md
+Task2: https://github.com/AlreadyBored/nodejs-assignments/blob/main/assignments/containerization-database-orm/assignment.md
 
 ## Prerequisites
 
 - Git - [Download & Install Git](https://git-scm.com/downloads).
 - Node.js - [Download & Install Node.js](https://nodejs.org/en/download/) and the npm package manager.
+- Docker desktop - [Download & Install Docker](https://www.docker.com/products/docker-desktop). Be sure to restart the computer.
 
 ## Downloading
 
@@ -29,6 +31,14 @@ npm start
 After starting the app on port (4000 as default) you can open
 in your browser OpenAPI documentation by typing http://localhost:4000/doc/.
 For more information about OpenAPI/Swagger please visit https://swagger.io/.
+
+## Running with docker
+
+```
+cp .env.example .env
+docker build -t arturkhelshtein/home-library-app:latest .
+docker run arturkhelshtein/home-library-app:latest
+```
 
 ## Testing
