+++ docker-compose.yml
@@ -1,15 +1,16 @@
 version: '3.8'
+
 services:
   postgres:
     image: postgres:15
     container_name: home-library-postgres
     restart: always
     ports:
-      - '5432:5432'
+      - '${POSTGRES_PORT}:5432'
     environment:
-      POSTGRES_USER: postgres
-      POSTGRES_PASSWORD: postgres
-      POSTGRES_DB: home_library
+      POSTGRES_USER: ${POSTGRES_USER}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD}
+      POSTGRES_DB: ${POSTGRES_DB}
     volumes:
       - pgdata:/var/lib/postgresql/data
     networks:
@@ -21,9 +22,11 @@ services:
       context: .
       dockerfile: Dockerfile
     ports:
-      - '4000:4000'
+      - '${PORT}:${PORT}'
     environment:
-      DATABASE_URL: postgres://postgres:postgres@postgres:5432/home_library
+      DATABASE_URL: ${DATABASE_URL}
+    env_file:
+      - .env
     depends_on:
       - postgres
     networks:
