+++ src/album/album.service.ts
@@ -4,7 +4,7 @@ import { EventEmitter2, OnEvent } from '@nestjs/event-emitter';
 import { CreateAlbumDto } from './dto/create-album.dto';
 import { UpdateAlbumDto } from './dto/update-album.dto';
 import { Album } from './entities/album.entity';
-import { albums } from './album.store';
+import { albums } from 'src/db';
 
 @Injectable()
 export class AlbumService {
