--- src/album/album.store.ts
@@ -1,3 +1,0 @@
-import { Album } from './entities/album.entity';
-
-export const albums: Album[] = [];
