+++ src/artist/artist.service.ts
@@ -4,7 +4,7 @@ import { EventEmitter2 } from '@nestjs/event-emitter';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { Artist } from './entities/artist.entity';
-import { artists } from './artist.store';
+import { artists } from 'src/db';
 
 @Injectable()
 export class ArtistService {
