--- src/artist/artist.store.ts
@@ -1,3 +1,0 @@
-import { Artist } from './entities/artist.entity';
-
-export const artists: Artist[] = [];
