--- src/favorites/favorites.store.ts
@@ -1,7 +1,0 @@
-import { Favorite } from './entities/favorite.entity';
-
-export const favorites: Favorite = {
-  tracks: [],
-  albums: [],
-  artists: [],
-};
