+++ src/track/track.service.ts
@@ -4,7 +4,7 @@ import { EventEmitter2, OnEvent } from '@nestjs/event-emitter';
 import { CreateTrackDto } from './dto/create-track.dto';
 import { UpdateTrackDto } from './dto/update-track.dto';
 import { Track } from './entities/track.entity';
-import { tracks } from './track.store';
+import { tracks } from 'src/db';
 
 @Injectable()
 export class TrackService {
