--- src/track/track.store.ts
@@ -1,3 +1,0 @@
-import { Track } from './entities/track.entity';
-
-export const tracks: Track[] = [];
