--- src/user/user.store.ts
@@ -1,3 +1,0 @@
-import { User } from './entities/user.entity';
-
-export const users: User[] = [];
